Fixed problems wherein the editor was not cancelled when the up and down arrow buttons were used essentially resulting in the value being applied to the wrong row.
Made changes to clear the values of the default row if the delete key is pressed or the remove menu item is used on the default row.

Made changes to prevent the Data view from refreshing for every change to the model. It is only refreshed when the operation is committed.

If a value is selected from the list of values for an expression, it is now quoted if the value is not a number.