- SCR(s) Resolved: bugzilla 103599 Resizing chart results in Eclipse hang up.
- Description:
When x-axis has a datetime type and non-category style. the AutoScale will try to find a prefered unit and step automatically to show the axis gracefully. the original algorithm uses an iteratively processing logic, but in some cases, it will enter infinite loop, hence the system hangs up. this has been fixed now.

- Regression ( Yes/No ):No

- Code Owner: Zhiqiang Qian

- Code Reviewers: SHA GUI

- Tests: Yes

- Test Automated (Yes/No, if "No" then explain why):No, gui feature

- Case Entries Resolved:
- Notes to Developers:
- Notes to QA:
- Notes to Documentation:
- Notes to Configuration Management:
- Notes to Support:
- Notes to Product Marketing: