- Summary: Add missing adapters in the Fill(background, outline color etc)
- Bugzilla Bug (s) Resolved:
122592

- Description:
Add missing adapters in the Fill(background, outline color etc) to ensure preview refreshed. This issue usually occurs when the EObject is null, and there's no adapter obtained from the constructor arguments. The standard fix is to pass the IWizardContext in all dialogs or other classes. It's not only to obtain the chart model, but also for feature extensibility.

- Tests Description:

- Notes to Build Team:

- Notes to Developers:

- Notes to QA:
Need to enable the transparency using the VM arguments " -DR31ENHANCE", and then scroll the slider in the color picker to see whether the preview is refreshed normally

- Notes to Documentation:

- Files Added:

- Files Edited: