- Summary: Chart Font properties should have "default" settings for style to be applied
- Bugzilla Bug (s) Resolved:
136584

- Description:
Added following features:
Font Type: replace Default by ¡°Auto¡±

Font Size: Change control to DropDown, first element being ¡°Auto¡±

Font Bold/Italic/Underline/Strikethrough: If selected, it will override the style. If unselected, we should remove the value from the model, so the style will be applied. This causes a problem when the style defines underline and you want to override it, but the designer UI has the same issue, I think it¡¯s ok  Let me know if you can think of something to resolve it. Also I think it makes sense to have the same buttons as in designer for those, particularly for Strikethrough, instead of showing a checkbox.

Font Alignment: Not sure if we support style for that, but if yes, we should have the ability to unselect: clicking on the selected item will unselect it, so that no alignment button is pushed. Btw, the buttons look a bit big to me, are those constrained by the size of the icons?

Font or Background Color: Replace the ¡°Reset Color¡± with ¡°Transparent¡± and add an ¡°Auto¡± button (selecting Transparent will show ¡°Transparent¡±, selecting ¡°Auto¡± will show ¡°Auto¡±)

Other colors (Series palette, shadows, border): Replace the ¡°Reset Color¡± with ¡°Transparent¡±, no need for Auto button as it¡¯s not supported and is confusing, particularly for Palette.


- Tests Description: Manual test
- Notes to Build Team: None
- Notes to Developers: None
- Notes to QA: None
- Notes to Documentation: None
- Files Added:

- Files Edited: