-Summary:Create the styles with the same name
- Bugzilla Bug (s) Resolved:136061

- Description:
I corrected the function GeneralPreferencePage.checkName(), currently if the style name is duplicate, it'll invoke an error dialog, not catch an exception.
But when creating a duplicate style in a a theme, if it throws an exception. the content of exception is not correct.
So reassign it to Model team now.

- Tests Description: Manual

- Notes to Build Team:

- Notes to Developers:

- Notes to QA:

- Notes to Documentation:

- Files Added:

- Files Edited: