- Summary:    Bugzilla Bug 146653Memory Leak in Report Designer
- Bugzilla Bug (s) Resolved:  146653

- Description:
We add a proxy edit part to support open files, which locates outside the workspace. The undo/redo action of styles text editor doesn't dispose because of this proxy is not the same insance of the multiple editor. Rewrite the eques method of this proxy editor to let these two action disposible.

- Tests Description  :
Manual test and Code review.

- Files Edited:

@changelist

- Files Added:

@addlist

- Files Deleted:

@deletelist

- Notes to Build Team:



- Notes to Developers:



- Notes to QA:



-          Notes to Documentation: