- Summary:  Bugzilla Bug 147020 Template reversion is not correct
- Bugzilla Bug (s) Resolved:  147020

- Description:
When we double click the template item, a copied data item created from the template item. At this time, the data item handle is not on the model tree. We can¡¯t get the parent¡¯s binding column. We need change the GUI behavior. Just convert it to a default value of the data item after user's first double click, then user can double click again to bring up the dialog.

- Tests Description  :   Manual Test

- Notes to Build Team:

- Notes to Developers:

- Notes to QA:

- Notes to Documentation:

- Files Edited:
@changelist
- Files Added:
@addlist
- Files Deleted
@deletelist