- Summary: cancel operation in hyperlink options (select a target report) is not user friendly
- Bugzilla Bug (s) Resolved:
149105


- Description:
Now, if user clicks Cancel button, the content in Text control will keep the same as before.

Fix another bug:
In linux envionment, when user open a FileDialog to select a file, the extension filter may be useless. So I check the file extension after the file is selected.
The update including BgImageFieldEditor, WizardSelectCssStylePage and HyperlinkBuilder.

- Tests Description  : Manual test

- Files Edited:

@changelist

- Files Added:

@addlist

- Files Deleted:

@deletelist

- Notes to Build Team:



- Notes to Developers:



- Notes to QA:



-  Notes to Documentation: