- Summary: Creating a New Sort or Filter for a Table -- Expression should initially be empty(Partial)
- Bugzilla Bug (s) Resolved:
 150579

- Description:
I add New/Edit Sortkey Dialog in this checkin.
The mockup of the other dialog,New/Edit FIlter Dialog will be attached later.

In attion, when using  New/Edit Sortkey Dialog, you may find that the context help is invalid, that's because the content has not beed added(I've added the help Id).
I'll assign Jane a bug to add context help for the two dialogs after this bug is fixed.


- Tests Description:
Manual test

- Files Edited:

@changelist

- Files Added:

@addlist

- Files Deleted:

@deletelist

- Notes to Build Team:



- Notes to Developers:



- Notes to QA:



-  Notes to Documentation: