- Summary: 116180-[SPM]Dataset can not be previewed in report designer on Mac
- Bugzilla Bug (s) Resolved: 116180

- Description:
This bug is caused by the org.eclipse.swt.widgets.Tree in Mac OS X. With the Tree's different implementaion in Mac, when user drag a selected object and move it in the source View, and if the View is a Tree, the method getSelection() of Tree will set the selected object as "null".
Our solution is add a private member "selectionList" into "DesignElementDragAdapter" class to store the selected object at first time, and set the drag source data with this selectionList. By this way, remove the impact from the swt Tree's implementation in Mac OS X.

- Tests Description: Manual test and code review.

- Notes to Build Team:

- Notes to Developers:

- Notes to QA:
This bug's fixing needs to tested in Windows, Linux and Mac OS X, to see if our modification in codes doesn't cause problem except for Mac. And we have tested it in Windows, it works well.

- Notes to Documentation:

- Files Added:
@addlist

- Files Edited:
@changelist