- Summary: Fix Bugzilla bug 152943-AggregateOn is always changed to "All" with focus change
- Bugzilla Bug (s) Resolved: 152943

- Description:
In previous codes, after putting cursor focus on expression, and if the expression which aggregrateon on group locates on table (group container is listing), it will set the "AggregateOn" as "null". Now, no matter where the expression locates (no matter the group container is listing or group) , it won't set the "Aggregateon" as "null".

- Tests Description: Manual test and code review.

- Notes to Build Team:

- Notes to Developers:

- Notes to QA:

- Notes to Documentation:

- Files Added:
@addlist

- Files Edited:
@changelist