Fix bugzilla bug 87077  Date parameter does not work for Actuate BIRT with IOD.
By default, the DateFormat parsing is lenient: If the input is not in the form used by this object's format method but can still be parsed as a date, then the parse succeeds.  Clients may insist on strict adherence to the format by calling setLenient(false).The call for method setLenient(false) is called for all DateFormat and SimpleDateFormat objects. And the bug can be fixed.