- Summary: Save thumnail into report design as embedded image
- Bugzilla Bug (s) Resolved:
 144843

- Description:
Add the feature:
Add a thumbnail button to general catogory of report property editor.
After clicking this button, it invokes a "Manage Thumbnail " Dialog. User can generate, remove or import an image as the thumbnail.

User was supposed to manage the thumbnail when publishing a template, but I don't add this feature this time, because of appearance problem. I don't know how to design the appearance to make it look better.

In addtion,it should have this feature: When saving a template, if the template to be saved has neither preview image nor thumbnail, the template will be saved with a thumbnail auto generated without telling the user.
But I don't implement it because of technical problem. I'll send myself a bug later to resolve this problem in future.

- Tests Description:
Manual test

- Files Edited:

@changelist

- Files Added:

@addlist

- Files Deleted:

@deletelist

- Notes to Build Team:



- Notes to Developers:



- Notes to QA:



-  Notes to Documentation: