- Summary: [Chart usability spec#2.3] UI refinement of interactivity dialog
- Bugzilla Bug (s) Resolved:
N/A

- Description:
1.	Quotes chart usability spec#2.3:
2.3.1   Auto saving for interactivity settings
Description: Currently, users have to click ¡°update¡± button to update interactivity settings. Users usually forget to click this so that they will lose the change. In V2.2, all settings in interactivity dialog will be saved automatically once they are changed.

2.3.2   Prevent adding several actions on the same event
Description: Currently, several actions on the same event are not supported in chart engine, so if users added one interactivity action and selected a event of the same type, Add button will be disabled.

2.	Refactored CustomChooserComposite, superclass of the custom combo. Now it supports Object as selected value.

- Tests Description: Manual test
- Notes to Build Team: None
- Notes to Developers: None
- Notes to QA: None
- Notes to Documentation: None

- Files Edited:

- Files Added:

- Files Deleted