- Summary: Jump to line of XML or report item when click on error in problem view
- Bugzilla Bug (s) Resolved:
 158509

- Description:
Clicking the problem view, GUI gets the information from the selected error item.
If this error item contains a visual report design element, then changes to layout or masterpage(If the element in the layout then switch to layout, else switch to masterpage) and select the element , if the error item doesn't contain an element information, then switch to xml source page, and jump to the error line.


- Tests Description:
Manual test

- Files Edited:

@changelist

- Files Added:

@addlist

- Files Deleted:

@deletelist

- Notes to Build Team:



- Notes to Developers:



- Notes to QA:



-  Notes to Documentation: