- Summary: Y Series type can not be set to Date Time for 3D Chart
 - Bugzilla Bug (s) Resolved:
161013

 - Description:
Cause: when chart engine computes the best scale of a 3D chart, there¡¯s a calculation error in loop, so it causes the BIRT not responding. This error is caused by a defect of the algorithm.
Solution: fix it by correcting the algorithm of computing if a Polygon contains the point.


- Tests Description:  Unit Test

- Notes to Build Team: N/A

- Notes to Developers: N/A

- Notes to QA:  N/A

- Notes to Documentation:  N/A

- Files Edited:

- Files Added:

- Files Deleted