Summary:A blemish appears when setting the width and height of chart legend to 0.
Bugzilla Bug (s) Resolved: 166012

Description:When rendering the 2D with Depth or 3D charts, the chart engine computes the "depth" in order to display the legend box as a cube instead of a square.

In fact, the blemish is the outline of the legend box. With the fix, the logic will skip the rendering part when the bounds of the legend box is 0.


Tests Description: Test Manually.

Files Edited: @changelist

Files Added: @addlist

Files Deleted: @deletelist

Notes to Build Team: None

Notes to Developers: None

Notes to QA: None

Notes to Documentation: None