Bugzilla Bug 172430 Chart will get smaller when you reopen the report containing a chart with blank height or width [0001]
Set default size for null chart dimension attributes when reopening report or cleaning in property editor. Now the chart with blank dimension will display the same with the default one in layout editor. But in preview, there may be size differences between them because of dpi issues, i.e. report engine wont explicitly set width/height for the chart without dimension attributes in preview, so its not recommended to leave blank for dimension attributes.