- Summary: [regression]Exception when new style in library theme.[01]
- Bugzilla Bug (s) Resolved:
 190114

- Description:
This is a regression bug.
Model changed the logic API ElementFactory.NewStyle().
In the API comment they said "Creates a new style element. The name is required. If the name is null, we will make a unique name for it.", so after the style is created, it should have a name, so GUI can set the name before check whether it's null.
But now, model changed the logic, it returns null, so GUI throws an exception, I'd add a check to judge if the string is null.

After I add the check, it won't throw excetpion. But the style doesn't have a default name, it's also a regression. I think model should handle this.
After all, model should not update the API before discussing with GUI, especially it's near release.

After adding code to check whether the name is null in GUI part, Re-assign this bug to model.

- Tests Description  :
Manual test

- Files Edited:


- Files Added:


- Files Deleted:


- Notes to Build Team:

- Notes to Developers:

- Notes to QA:

-  Notes to Documentation: