- Summary: Fix two bugs:198774 and 198432
- Bugzilla Bug (s) Resolved:
198774 and 198432

- Description:
198774:Can not publish to template folder when set the preview image as thnmbnail image[00]
198432:Space is missing in org.eclipse.birt.report.designer.core message

About bug 198774,
In the old version(before 2.2 M6). preview image name presents the path of image filename, but Thumbnail is an image stored in the report file.
But now, if thumbnail exist, then the image name is the name of thumbnail.

About the spec change, please reference bug 144843.

Now,
If thumbnail image exists, it won't check whether the file exists or not.
If thumbnail image doesn't exist, and the preview image name is not empty, then it'll check if the file exists or not.

Fixed.

- Tests Description  :
Manual test

- Files Edited:


- Files Added:


- Files Deleted:


- Notes to Build Team:

- Notes to Developers:

- Notes to QA:

-  Notes to Documentation: