- Summary: Parameter dialog issue when change parameter type.[0705]
- Bugzilla Bug (s) Resolved:
204518

- Description:
In step 2, when changing type to Date, value 0 is invalid, so it's removed, and user can only see value 1 in the list.
But the default value is still 0.
When user clicks OK button, the program validates default value, and it throws an exception.

So I think when changing data type, program should also check the default value at the same time. If it's invalid, remove it, or keep it unchanged.

Fixed.

- Tests Description  :
Manual test

- Files Edited:


- Files Added:


- Files Deleted:


- Notes to Build Team:

- Notes to Developers:

- Notes to QA:

-  Notes to Documentation: