- Summary:  [Linux-WS4]The size of filter issue[00]
- Bugzilla Bug (s) Resolved:
202567

- Description:
Use Combo instead of ExpressionValue.
As before, ExpressionValue is composed of a Text and a Button. In Linux enviroment, it looks weird.
So use Combo control to take the place of our custom composite(ExpressionValue) to improve the aspect.
Update all the ExpressionValue instances in dialog FilterConditionBuilder.
Fixed.

- Tests Description  :
Manual test

- Files Edited:


- Files Added:


- Files Deleted:


- Notes to Build Team:

- Notes to Developers:

- Notes to QA:

-  Notes to Documentation: