- Summary: [Accessibility] Some issues in properties editor shortcuts.
- Bugzilla Bug (s) Resolved:
231273

- Description:
In fact, shortcut clash is avoidless because of the characters are limited, what we can do is to minimize the possibility. I've updated the shortcuts as following:
    Layout Preference: -> Lay&out Preference:
    &Resource file: -> Res&ource file:
on border page:  &St&yle: -> Style:
 report page: Path: -> &Path:
 library page: Namespace: -> &Namespace:
In item 4, the text is disabled, so you cannot set focus. It can be focused if it's enabled, even if it's read-only.


- Tests Description  :
Manual test


- Files Edited:



- Files Added:



- Files Deleted:



- Notes to Build Team:


- Notes to Developers:


- Notes to QA:


-  Notes to Documentation: