- Summary: Running aggregate without no grouping has error.
 - Bugzilla Bug (s) Resolved: 246810

 - Description:
The issue is caused by error checking of aggregate on value series. The running aggregate on value series does not depend on if category series is set default group, it should check aggregate setting on value series fist. And when checking if chart model is set aggregate, it should iterate all value series definition.

- Tests Description:  Manual Test

- Notes to Build Team: N/A

- Notes to Developers: N/A

- Notes to QA:  N/A

- Notes to Documentation:  N/A

- Files Edited:


- Files Added:


- Files Deleted