- Summary: TED 11045, Grouping Values (Y Axis) returns wrong data in the chart
 - TED Bug (s) Resolved: 11045

 - Description:
The issue is caused by an error in checking if category expression or y optional expression is a group expression, the reason is that the y optional expression is a script and has comments.If the expression is a script, it is failure to validate the grouping case. Now the checking logic was fixed to check script as grouping expression.
Another minor issue also was fixed. When chart inherit table, the new binding also can be added on chart item and the binding name is allowed duplicate with container's. While opening chart builder and do live preview, chart will get all bindings from chart item and its contains to execute query, in older version(5/27/2008), DTE will ignore duplicate bindings, but now the logic was changed by DTE and didn't allow duplicate bindings, DTE will throw exceptions and fail to do query. To fixed the issue, chart now filters container's duplicate bindings.

- Tests Description:  Manual Test

- Notes to Build Team: N/A

- Notes to Developers: N/A

- Notes to QA:  N/A

- Notes to Documentation:  N/A

- Files Edited:


- Files Added:


- Files Deleted