- Summary: TED7909, Chinese characters in chart can not be displayed in iPortal which is installed on Linux.
 - TED Bug (s) Resolved: 7909

 - Description:
 Checked in Head & 2.3.2Branch.
The reason is we use java.awt.Font to wrap current font setting and call Font.getFamily() as the font family name in SVG document. If current daemon is running on linux, the calling of Font.getFamily() might not get valid font family and a default 'Dialog' font family will be as current SVG font family name. it caused client didn't use correct font to render.
Since the SVG document is rendered in client, so the font setting in chart model should be directly set into SVG document as font family name when current output format is SVG. Now fixed it.

- Tests Description:  Manual Test

- Notes to Build Team: N/A

- Notes to Developers: N/A

- Notes to QA:  N/A

- Notes to Documentation:  N/A

- Files Edited:


- Files Added:


- Files Deleted